+++ spinta/utils/url.py
@@ -1,6 +1,8 @@
 import urllib.parse
+from typing import List
 
 from spinta import exceptions
+from spinta.components import UrlParseNode
 from spinta.spyna import unparse
 
 RULES = {
@@ -40,28 +42,33 @@ RULES = {
 }
 
 
-def apply_query_rules(RULES, params):
+def apply_query_rules(rules, params):
     for param in params:
         name = param['name']
         args = param['args']
 
-        if name not in RULES:
+        if name not in rules:
             raise exceptions.UnknownRequestParameter(name=name)
 
-        rules = RULES[name]
-        maxargs = rules.get('maxargs')
-        minargs = rules.get('minargs', 0 if maxargs == 0 else 1)
+        rule = rules[name]
+        maxargs = rule.get('maxargs')
+        minargs = rule.get('minargs', 0 if maxargs == 0 else 1)
         if len(args) < minargs:
             # FIXME: This should be an UserError.
-            raise Exception(f"At least {minargs} argument is required for {name!r} URL parameter.")
+            raise Exception(
+                f"At least {minargs} argument is required for {name!r} URL "
+                "parameter."
+            )
 
         if maxargs is not None and len(args) > maxargs:
             # FIXME: This should be an UserError.
-            raise Exception(f"URL parameter {name!r} can only have {maxargs} arguments.")
+            raise Exception(
+                f"URL parameter {name!r} can only have {maxargs} arguments."
+            )
     return params
 
 
-def parse_url_path(path):
+def parse_url_path(path) -> List[UrlParseNode]:
     query = []
     name = None if path.startswith(':') else 'path'
     args = []
@@ -85,7 +92,7 @@ def parse_url_path(path):
     return query
 
 
-def build_url_path(query):
+def build_url_path(query: List[UrlParseNode]):
     parts = []
     other = []
     for param in query: